refactored files to react app scope
