--- App.css
@@ -1,69 +1,0 @@
-/* 
-* This file holds the CSS commentary for App.js
-* It sets the style guidelines for the Log In/Sign Up page.
-*
-* Contributors: Tabassum Alam
-*/
-
-.App {
-  text-align: center;
-}
-
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.App-link {
-  color: rgb(103, 211, 103);
-}
-
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
-
-/* puts Log In and Sign Up pages side by side */
-.main-display{
-  display:flex;
-  flex-direction: row;
-  margin: 0 auto;
-  margin-left: 10%;
-  margin-right: 10%;
-  box-align: center;
-  position: relative;
-}
-
-.image-display {
-  margin-top: 10%;
-  display: flex;
-  position: relative;
-}
-
-.tab-display {
-  display: inline;
-  position: relative;
-  margin: 0 auto;
-}
-
