+++ client/src/App.js
@@ -8,22 +8,27 @@
 import React from 'react';
 import './App.css';
 import Header from './components/Header';
-import LogIn from './components/LogIn';
-import SignUp from './components/SignUp';
-import LeftSideBar from "./components/LeftSideBar";
-
+import MainLogInSignUp from './components/MainLogInSignUp';
+import hi_anim from './media/hi_anim.mp4';
+import hi from './media/hi.jpg';
 
 // function to create the Log In/Sign Up page
 function App() {
+
   return (
     <div>
-      {/*<Header />*/}
+      <Header />
       {/* LogIn and SignUp will appear side by side */}
       <div className="main-display">
-        {/*<LogIn />*/}
-        <LeftSideBar/>
-        {/*<SignUp />*/}
-      </div>
+        <div className="image-display">
+          <video className="image-display" id="welcome-video" muted autoPlay poster={hi} loop 
+          src={hi_anim} width="400" height="400">
+          </video> 
+        </div>
+        <div className="tab-display">  
+          <MainLogInSignUp />
+        </div>  
+    </div>
     </div>
   )
 }
