--- index.css
@@ -1,68 +1,0 @@
-/* ---------------------------------------------------------------------
-Contributors: Darien
-Overriding styles for every page
---------------------------------------------------------------------- */
-@import url('https://fonts.googleapis.com/css?family=PT+Sans&display=swap');
-
-/* ---------------------------------------------------------------------
-Colors
---------------------------------------------------------------------- */
-:root{
-  --bg-green: rgb(103, 211, 103);
-  --bg-light-green: rgb(162, 226, 162);
-  --bg-white: #eeeeee;
-  --button-green: rgb(23, 82, 23);
-  --button-hover: rgb(30, 107, 30);
-  --button-active: rgb(19, 68, 19);
-}
-
-
-/* ---------------------------------------------------------------------
-Override Scrollbar
---------------------------------------------------------------------- */
-*{
-  margin: 0;
-  padding: 0;
-  font-family: 'PT Sans', sans-serif;
-}
-
-body{
-  width: 100vw;
-  height: 100vh;
-}
-
-a{
-  color: black;
-  text-decoration: none;
-  font-style: none;
-  white-space: nowrap;
-  cursor: pointer;
-  display: block;
-}
-
-
-/* ---------------------------------------------------------------------
-Override Scrollbar
---------------------------------------------------------------------- */
-::-webkit-scrollbar-track{
-  -webkit-box-shadow: inset 0 0 6px rgba(255, 255, 255, 0.3);
-  background-color: rgba(255, 255, 255, 0);
-}
-
-::-webkit-scrollbar{
-  width: 5px;
-  height: 3px;
-  background-color: rgb(255, 255, 255);
-  cursor: pointer;  
-}
-
-::-webkit-scrollbar-thumb{
-  background-color: #28bd3f;
-}
-
-
-/* ---------------------------------------------------------------------
-Override Highlight
---------------------------------------------------------------------- */
-::-moz-selection {background: #92f8a0;}
-::selection {background: #92f8a0;}
